chore: move calculateTxBalance to the transaction util
