+++ src/types.ts
@@ -331,7 +331,6 @@ export interface IStorage {
   // UTXOs
   getAllUtxos(): AsyncGenerator<IUtxo>;
   selectUtxos(options: Omit<IUtxoFilterOptions, 'reward_lock'>): AsyncGenerator<IUtxo>;
-  calculateTxBalance(tx: IDataTx): Promise<Map<string, Record<'funds'|'mint'|'melt', number>>>;
   fillTx(tx: IDataTx, options: IFillTxOptions): Promise<{inputs: IDataInput[], outputs: IDataOutput[]}>;
   utxoSelectAsInput(utxo: IUtxoId, markAs: boolean, ttl?: number): Promise<void>;
